refactor(ui): clean up the Markdown component code
